Add start and end date
Addresses Google FO: This charter does not have an end date on it; no open-ended charters should be approved. We would suggest a two-year (from start) charter, which would remove this FO.

Proposed response: We set the expected start date of the charter to 1 August 2021 and the end date to 31 July 2023, making this a 2-year charter.